feat(telegram): Add subscription api
Refactor Telegram client so it is couples into namespaces based on the
api purpose. We now have chats, updates and payments. This is basically
refactoring to make sure we stop writing huge JS modules.

in the addition, we switched to use ISO to Locale helper to ensure we do
not have the copies of the same logic in many places
