+++ e2e/dm.ru.spec.ts
@@ -20,7 +20,7 @@ import {
 } from "./helpers/dependencies.js";
 import { mockTableCreation, Pool as MockPool } from "../src/db/__mocks__/pg.js";
 import { getSupportedAudioFormats } from "../src/text/utils.js";
-import type { TgChatType } from "../src/telegram/api/types.js";
+import type { TgChatType } from "../src/telegram/api/groups/chats/chats-types.js";
 import type { LanguageCode } from "../src/recognition/types.js";
 import type { VoidPromise } from "../src/common/types.js";
 
@@ -110,7 +110,7 @@ describe("[russian language]", () => {
     const getDb = init.getDb;
     const localhostUrl = init.localhostUrl;
     const TelegramBotModel = init.TelegramBotModel;
-    const TelegramApi = init.TelegramApi;
+    const TelegramBaseApi = init.TelegramBaseApi;
     const mockTgGetWebHook = initTest.mockTgGetWebHook;
     const mockTgSetWebHook = initTest.mockTgSetWebHook;
     const mockTgSetCommands = initTest.mockTgSetCommands;
@@ -142,7 +142,7 @@ describe("[russian language]", () => {
     bot = new TelegramBotModel("telegram-api-token", converter, db);
     bot.setHostLocation(hostUrl, launchTime);
 
-    telegramServer = nock(TelegramApi.url);
+    telegramServer = nock(TelegramBaseApi.url);
     host = request(hostUrl);
 
     testLangId = "ru-RU";
