+++ e2e/lifecycle.spec.ts
@@ -64,7 +64,7 @@ describe("[lifecycle]", () => {
     const getDb = init.getDb;
     const TelegramBotModel = init.TelegramBotModel;
     const launchTime = init.launchTime;
-    const TelegramApi = init.TelegramApi;
+    const TelegramBaseApi = init.TelegramBaseApi;
 
     mockGoogleAuth();
 
@@ -89,7 +89,7 @@ describe("[lifecycle]", () => {
     bot = new TelegramBotModel("telegram-api-token", converter, db);
     bot.setHostLocation(hostUrl, launchTime);
 
-    telegramServer = nock(TelegramApi.url);
+    telegramServer = nock(TelegramBaseApi.url);
     host = request(`${localhostUrl}:${appPort}`);
   });
